refactor(log): refactor Logger so it's testable
