+++ src/config/config.ts.example
@@ -1,5 +1,3 @@
-import { LogLevel } from '../lib/logger'
-
 /**
  * Enable this if you want a lot of text to be logged to console.
  * @type {boolean}
