--- src/lib/logger/logLevel.ts
@@ -1,6 +1,0 @@
-export enum LogLevel {
-  ERROR,
-  WARNING,
-  INFO,
-  DEBUG
-}
