+++ src/lib/logger/logger.g.d.ts
@@ -1,0 +1,11 @@
+/**
+ * Available log levels for the logger library.
+ *
+ * @enum {number}
+ */
+declare const enum LogLevel {
+  ERROR,
+  WARNING,
+  INFO,
+  DEBUG
+}
